requirements/app.txt
streamlit==1.56.0
ultralytics==8.4.37
torch==2.11.0
torchvision==0.26.0
opencv-python-headless==4.13.0.92
numpy==2.4.4
Pillow==12.2.0
PyYAML==6.0.3
